separate test classes for AddAuditLogUseCase and GetAuditLogUseCase
